refactor(sim): split consumer downgrade facade
