+++ crates/assay-sim/src/attacks/consumer_downgrade_next/types.rs
@@ -1,0 +1,23 @@
+use serde::Serialize;
+
+#[derive(Debug, Clone, Serialize)]
+pub struct ConsumerResult {
+    pub vector_id: String,
+    pub condition: String,
+    pub realism_class: String,
+    pub canonical_classification: String,
+    pub consumer_classification: String,
+    pub downgrade_occurred: bool,
+    pub outcome: ConsumerOutcome,
+    pub hypothesis_tags: Vec<String>,
+}
+
+#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum ConsumerOutcome {
+    NoEffect,
+    RetainedNoDowngrade,
+    DowngradeWithCorrectDetection,
+    SilentDowngrade,
+    SilentTrustUpgrade,
+}
